% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIest.R
\name{AIest}
\alias{AIest}
\title{Attraction Index}
\usage{
AIest(
  pp,
  correction = "Ripley",
  dim = NULL,
  dim_lims = NULL,
  dr = NULL,
  rmax = NULL,
  nlarge = 3000
)
}
\arguments{
\item{pp}{The observed point pattern, from which an estimate of AI(r) will be computed.
An object of class "ppp", or data in any format acceptable to as.ppp().}

\item{correction}{Optional. A character vector containing one and only one of
the options "none", "border", "bord.modif", "isotropic", "Ripley", "translate",
"translation", "rigid", "none", "periodic", "good" or "best".
It specifies the edge correction to be applied. Note that the option "all"
or providing multiple edge correction methods is not supported due to
performance reasons. Defaults to "Ripley".}

\item{dim}{Optional. The dimension of the basis used to represent the smooth term within
the scam model formula. If not provided, the rule of thumb is used, i.e.
dim = sqrt(number of points).}

\item{dim_lims}{Optional. The integer vector with 2 values: the lower and
upper limits of the possible value of dim, c(lower, upper). lower > upper
is not allowed.}

\item{dr}{Optional. The delta r to compute the derivative of the estimated
number of points. If not provided a rule of thumb is supplied, i.e. 1/5 of
the distance between subsequent r values in the K-function estimate is used.}

\item{rmax}{Optional. Maximum desired value of the argument.}

\item{nlarge}{Optional. Efficiency threshold. If the number of points exceeds
nlarge, then only the border correction will be computed (by default),
using a fast algorithm.}
}
\value{
An object of class "fv", see fv.object, which can be plotted directly using plot.fv.
}
\description{
Attraction Index
}
\examples{
#TBD
}
