% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCFest.R
\name{choose_basis_dim}
\alias{choose_basis_dim}
\title{The number of B-splines to use in the scam model}
\usage{
choose_basis_dim(sample_size, dim_lims = NULL)
}
\arguments{
\item{sample_size}{Number of points in a point pattern}

\item{dim_lims}{Optional. The integer vector with 2 values: the lower and
upper limits of the possible value of dim, c(lower, upper). lower > upper
is not allowed.}
}
\value{
An integer value that represents the number of B-splines to use
in a scam model. The rule of thumb dim = sqrt(sample_size) is used and
if dim_lims is provided it is used to clip the value
}
\description{
The number of B-splines to use in the scam model
}
