% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIest.R
\name{compute_ai}
\alias{compute_ai}
\title{The AI value computation}
\usage{
compute_ai(r, pn, pn_deriv, ai_lims = c(-1, 1))
}
\arguments{
\item{r}{A vector of distances}

\item{pn}{A vector of estimated expected number of points within a distance r}

\item{pn_deriv}{A vector of estimated derivative of the expected number of points
within a distance r}

\item{ai_lims}{Optional. The double vector with 2 values: the lower and
upper limits of the AI, c(lower, upper). The lower value corresponds to the AI value
for maximal dispersion, the upper to the AI value for maximal clustering.
lower > upper is not allowed.}
}
\value{
A vector with AI estimate at r
}
\description{
The AI value computation
}
