% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCFest.R
\name{compute_lcf}
\alias{compute_lcf}
\title{The LCF value computation}
\usage{
compute_lcf(r, pn, pn_deriv, lcf_lims = c(-1, 1))
}
\arguments{
\item{r}{A vector of distances}

\item{pn}{A vector of estimated expected number of points within a distance r}

\item{pn_deriv}{A vector of estimated derivative of the expected number of points
within a distance r}

\item{lcf_lims}{Optional. The double vector with 2 values: the lower and
upper limits of the LCF, c(lower, upper). The lower value corresponds to the LCF value
for maximal dispersion, the upper to the LCF value for maximal clustering.
lower > upper is not allowed.}
}
\value{
A vector with LCF estimate at r
}
\description{
The LCF value computation
}
