% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCFest.R
\name{LCF}
\alias{LCF}
\title{LCF estimate}
\usage{
LCF(pn_est_df, r = NULL, dim, est_name)
}
\arguments{
\item{pn_est_df}{The data frame with estimated number of points at distance r
that will be used to estimate LCF. Should have two columns
"r" (distance) and "pn" (estimated number of points)}

\item{r}{Optional. Optional. Vector of values for the argument r at which LCF(r)
should be evaluated. The values must be in increasing order. Advanced use only.}

\item{dim}{The dimension of the basis used to represent the smooth term within
the scam model formula. If not provided, the rule of thumb is used, i.e.
dim = sqrt(number of points).}

\item{est_name}{The name of the column that contains the LCF's empirical estimate.}
}
\value{
A data frame with 3 columns: r contains distances, theo contains LCF's
theoretical value at the corresponding distance and columns names with est_name
contains the empirical estimate of the function
}
\description{
Estimates LCF by using smooth approximation of the empirical estimate of
the number of points, N(r), and computing its derivative
}
