% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIest.R
\name{single_mpi_derivative}
\alias{single_mpi_derivative}
\title{Derivative of a scam model with a single MPI term}
\usage{
single_mpi_derivative(model, data)
}
\arguments{
\item{model}{A scam model with a single monotone increasing P-spline (MPI)
term}

\item{data}{A data frame containing the values of the named covariates
at which the smooth term is to be evaluated.}
}
\value{
A vector that contains the derivative of the given scam model at
the valuse provided with the data parameter
}
\description{
Derivative of a scam model with a single MPI term
}
