% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCFest.R
\name{LCFcross}
\alias{LCFcross}
\title{Multitype LCF (Cross-type)}
\usage{
LCFcross(
  pp,
  i,
  j,
  correction = "Ripley",
  r = NULL,
  dim = NULL,
  dim_lims = NULL,
  rmax = NULL
)
}
\arguments{
\item{pp}{The observed point pattern, from which an estimate of the cross-LCF will be computed.
It must be a multitype point pattern (a marked point pattern whose marks are a factor).}

\item{i}{The type (mark value) of the points in pp from which distances are measured.
Must be a character string. Defaults to the first level of marks(pp).}

\item{j}{The type (mark value) of the points in pp to which distances are measured.
Must be a character string. Defaults to the second level of marks(pp).}

\item{correction}{Optional. A string containing one of
the options "none", "border", "bord.modif", "isotropic", "Ripley", "translate",
"translation", "rigid", "none", "periodic", "good" or "best".
It specifies the edge correction to be applied. Note that the option "all"
or providing multiple edge correction methods is not supported due to
performance reasons. Defaults to "Ripley".}

\item{r}{Optional. Vector of values for the argument r at which cross-LCF(r)
should be evaluated. The values must be in increasing order. Advanced use only.}

\item{dim}{Optional. The dimension of the basis used to represent the smooth term within
the scam model formula. If not provided, the rule of thumb is used.}

\item{dim_lims}{Optional. The integer vector with 2 values: the lower and
upper limits of the possible value of dim, c(lower, upper). lower > upper
is not allowed.}

\item{rmax}{Optional. Maximum desired value of the argument.}
}
\value{
An object of class "lcffv", inherited from fv.object, which can be
plotted directly using plot.lcffv. Essentially a data frame that contains
columns

r      the value of the argument r at which cross-LCF(r) is estimated
theo   the theoretical value of cross-LCF(r) for a marked Poisson process, 0

And a column with the empirical estimate of cross-LCF(r) obtained from the point pattern
using the specified edge correction method.
}
\description{
For a multitype point pattern, estimate the cross-LCF which estimates the spatial
distribution of objects of type j with respect to objects of type i.
}
\examples{

# TODO: Expand this

library(spatstat.data)

lcf_cross <- LCFcross(amacrine, "on", "off")
plot(lcf_cross)

}
