% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIest.R
\name{plot.aifv}
\alias{plot.aifv}
\title{Plot Function Value for AI}
\usage{
\method{plot}{aifv}(x, main = "", ...)
}
\arguments{
\item{x}{An object of the class "aifv" that contains the variables to be
plotted.}

\item{main}{A title of the plot.}

\item{...}{Extra arguments passed to the spatstat.explore::plot.fv}
}
\value{
Invisible: either NULL, or a data frame giving the meaning of the
different line types and colours.
}
\description{
Plot Function Value for AI
}
\examples{
library(spatstat.random)

rpp <- rpoispp(500)
ai_rand <- AIest(rpp)
plot(ai_rand, main = "AI for a random point pattern")

}
