% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCFest.R
\name{plot.lcffv}
\alias{plot.lcffv}
\title{Plot Function Value for LCF}
\usage{
\method{plot}{lcffv}(x, ylim = c(-1, 1), main = "", ...)
}
\arguments{
\item{x}{An object of the class "lcffv" that contains the variables to be
plotted.}

\item{ylim}{(optional) range of y axis. Default is set to the lower and
upper limits of the LCF.}

\item{main}{(optional) A title of the plot.}

\item{...}{Extra arguments passed to the spatstat.explore::plot.fv}
}
\value{
Invisible: either NULL, or a data frame giving the meaning of the
different line types and colours.
}
\description{
Plot Function Value for LCF
}
\examples{
library(spatstat.random)

rpp <- rpoispp(500)
lcf_rand <- LCFest(rpp)
plot(lcf_rand, main = "LCF for a random point pattern")

}
